quiz-api/requirements.txt
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
PyJWT==2.5.0
six==1.16.0
Werkzeug==2.2.3
